Added in styles
Added in all of the links for the stylesheets on all of the pages. Renamed all of the files to remove the _txt part of the name. Added a flex box to the body. Set the scaling size of the body flex items. Added in media queries to allow mobile support.